refactor: clean controller code, feature: add loggin middleware
